(refactor) AdminMusicService, repository refactoring
