+++ src/main/java/com/grepp/moodlink/app/model/data/music/MusicService.java
@@ -1,17 +1,12 @@
 package com.grepp.moodlink.app.model.data.music;
 
-import com.grepp.moodlink.app.model.data.book.entity.Book;
 import com.grepp.moodlink.app.model.data.music.dto.MusicDto;
 import com.grepp.moodlink.app.model.data.music.entity.Music;
-import com.grepp.moodlink.infra.error.exceptions.CommonException;
 import com.grepp.moodlink.infra.imgbb.ImgUploadTemplate;
-import com.grepp.moodlink.infra.response.ResponseCode;
-import java.io.IOException;
 import java.time.LocalDate;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Optional;
-import java.util.UUID;
 import java.util.regex.Matcher;
 import java.util.regex.Pattern;
 import java.util.stream.Collectors;
@@ -19,11 +14,7 @@ import java.util.stream.Collectors;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.modelmapper.ModelMapper;
-import org.springframework.data.domain.Page;
-import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
-import org.springframework.transaction.annotation.Transactional;
-import org.springframework.web.multipart.MultipartFile;
 
 @Service
 @RequiredArgsConstructor
@@ -31,8 +22,6 @@ import org.springframework.web.multipart.MultipartFile;
 public class MusicService {
 
     private final MusicRepository musicRepository;
-    private final ModelMapper mapper;
-    private final ImgUploadTemplate imgUploadTemplate;
 
     public void saveMusic(List<MusicDto> musicDtos) {
 
@@ -52,71 +41,6 @@ public class MusicService {
 
             musicRepository.save(music);
         }
-    }
-
-    public Page<MusicDto> findPaged(Pageable pageable) {
-        return musicRepository.findPaged(pageable)
-            .map(e -> mapper.map(e, MusicDto.class));
-    }
-
-    @Transactional
-    public void addMusic(List<MultipartFile> thumbnail, MusicDto dto) {
-
-        if(musicRepository.existsByTitleAndSinger(dto.getTitle(),dto.getSinger()))
-            throw new CommonException(ResponseCode.DUPLICATED_DATA);
-
-        try {
-
-            uploadImage(thumbnail, dto);
-            Music music = mapper.map(dto, Music.class);
-
-            long count = musicRepository.count();
-            music.setId("S"+count);
-
-            log.info("{}",music);
-
-            musicRepository.save(music);
-        } catch (IOException e) {
-            throw new CommonException(ResponseCode.INTERNAL_SERVER_ERROR, e);
-        }
-    }
-
-    private void uploadImage(List<MultipartFile> thumbnail, MusicDto dto) throws IOException {
-        if(thumbnail != null){
-            MultipartFile file =  thumbnail.getFirst();
-            String originFileName = file.getOriginalFilename();
-            if (originFileName != null && originFileName.contains(".")) {
-                String ext = originFileName.substring(originFileName.lastIndexOf("."));
-                String renameFileName = UUID.randomUUID().toString() + ext;
-
-                String thumbnailUrl = imgUploadTemplate.uploadImage(file, renameFileName);
-                dto.setThumbnail(thumbnailUrl);
-            }
-        }
-    }
-
-    public MusicDto findById(String id) {
-        return musicRepository.findById(id).map(e -> mapper.map(e, MusicDto.class)).orElse(null);
-    }
-
-    public void updateMusic(List<MultipartFile> thumbnail, MusicDto dto) {
-
-        try {
-           uploadImage(thumbnail,dto);
-            // 업데이트
-            musicRepository.updateBook(dto);
-
-            log.info("{}",dto);
-
-
-        } catch (IOException e) {
-            throw new CommonException(ResponseCode.INTERNAL_SERVER_ERROR, e);
-        }
-    }
-
-    @Transactional
-    public void deleteMusic(String id) {
-        musicRepository.findById(id).ifPresent(Music::unActivated);
     }
 
     public List<String> parseRecommend(String musicResult) {
